+++ CHANGELOG.md
@@ -27,9 +27,13 @@ To publish a new release run `scripts/release` from the project directory.
     - Reduce error type of `mdcat::push_tty` to `std::io::Error`: `mdcat` never
       visibly fails unless it really can’t write to the output stream.
 
+### Fixed
+- Respect `--local-only` and resource access policy; this got lost in some refactoring (see [GH-146]).
+
 [GH-142]: https://github.com/lunaryorn/mdcat/issues/142
 [GH-139]: https://github.com/lunaryorn/mdcat/issues/139
 [anyhow]: https://docs.rs/crate/anyhow
+[GH-146]: https://github.com/lunaryorn/mdcat/issues/146
 
 ## [0.17.1] – 2020-05-24
 ### Fixed